requirements.txt
# --- Core Bot Framework ---
aiogram>=3.4.1
uvloop>=0.19.0; sys_platform == 'linux'

# --- Database & Migrations ---
sqlalchemy[asyncio]>=2.0.27
asyncpg>=0.29.0
alembic>=1.13.1

# --- Task Queue & Scheduling ---
celery[redis]>=5.3.6
redis>=5.0.1

# --- AI & External APIs ---
google-generativeai>=0.4.1
aiohttp>=3.9.3

# --- Config & Logging ---
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
structlog>=24.1.0

# --- Image Processing ---
pillow>=10.2.0

# --- Utilities ---
pydantic>=2.6.3
